[KAN-29] rename: 로그인/회원가입 컴포넌트 폴더구조 변경
