+++ src/app/(auth)/_component/Footer.tsx
