+++ src/app/(auth)/_component/FormField.tsx
@@ -1,5 +1,5 @@
-import Input from '../Input/Input';
-import PasswordInput from '../Input/PasswordInput';
+import Input from '../../components/Input/Input';
+import PasswordInput from '../../components/Input/PasswordInput';
 import { errorMessageStyle } from './FormOptions';
 
 interface SignUpData {
